style: clear ruff findings in the files this branch owns
RUF046 (redundant int(round(...))) in estimate.py, UP035 / FURB188 / PLW1510 in
judge.py, and the unused pathlib import plus import order in tools.py.

The lint gate is still red for reasons that predate this branch and live in files
another agent owns: UP035 in sources.py, store.py, turns.py, ledger.py and
lenses/__init__.py, FURB188 in store.py, SIM102 in turns.py, I001 in __init__.py.
main is red for exactly those.

+++ astern/estimate.py
@@ -281,7 +281,7 @@ def predict(model: dict, features: dict) -> dict:
     y_out = max(0.0, float(model["output"]["mean"]))
     total = y_in + y_out
     band = float(inp["resid_std"] or 0.0) + float(model["output"]["std"] or 0.0)
-    return {"input_tokens": int(round(y_in)), "output_tokens": int(round(y_out)),
-            "total_tokens": int(round(total)),
-            "low": int(round(max(0.0, total - band))), "high": int(round(total + band)),
+    return {"input_tokens": round(y_in), "output_tokens": round(y_out),
+            "total_tokens": round(total),
+            "low": round(max(0.0, total - band)), "high": round(total + band),
             "cost_usd": round(total * model.get("cost_per_total_token", 0.0), 6)}
